requirement.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
bcrypt==3.2.0
python-multipart
python-jose
email-validator